refactor(plugins): extract exceptions.py and rename types.py to schemas.py
Move GoogleDriveAPIError out of client.py into a dedicated exceptions.py.
client.py re-exports it via __all__ for backward compatibility. Rename
types.py to schemas.py since the file contains Pydantic request/response
models, not generic types.
