+++ app/core_plugins/googledrive/client.py
@@ -9,13 +9,9 @@ from typing import Any, Optional, Type
 import aiohttp
 from aiohttp import MultipartWriter
 
+from app.core_plugins.googledrive.exceptions import GoogleDriveAPIError
 
-class GoogleDriveAPIError(RuntimeError):
-    """Raised when a Google Drive API request fails."""
-
-    def __init__(self, status_code: int, message: str):
-        self.status_code = status_code
-        super().__init__(f"Google Drive API error ({status_code}): {message}")
+__all__ = ["GoogleDriveAPIError", "GoogleDriveClient"]
 
 
 class GoogleDriveClient:
