+++ app/core_plugins/googledrive/routes/files.py
@@ -17,7 +17,7 @@ from app.core_plugins.googledrive.constants import DRIVE_MAX_UPLOAD_BYTES
 from app.core_plugins.googledrive.oauth import get_valid_access_token
 from app.core_plugins.googledrive.routes.dependencies import require_user_id
 from app.core_plugins.googledrive.routes.route_utils import get_drive_credentials
-from app.core_plugins.googledrive.types import (
+from app.core_plugins.googledrive.schemas import (
     DriveBrowseItem,
     DriveBrowseResponse,
     DriveFileResponse,
