+++ app/core_plugins/googledrive/routes/folders.py
@@ -12,7 +12,7 @@ from app.core_plugins.googledrive.client import GoogleDriveClient
 from app.core_plugins.googledrive.oauth import get_valid_access_token
 from app.core_plugins.googledrive.routes.dependencies import require_user_id
 from app.core_plugins.googledrive.routes.route_utils import get_drive_credentials
-from app.core_plugins.googledrive.types import (
+from app.core_plugins.googledrive.schemas import (
     CreateFolderRequest,
     DriveFileResponse,
     DriveFolderResponse,
